doing todo 4

+++ Dev_class9_handout/Motor2D/j1PerfTimer.cpp
@@ -26,7 +26,7 @@ void j1PerfTimer::Start()
 // ---------------------------------------------
 double j1PerfTimer::ReadMs() const
 {
-	return (double)((double)(SDL_GetPerformanceCounter() - started_at) * 1000 / (double) frequency);
+	return ((double)(SDL_GetPerformanceCounter() - started_at) * 1000 /  frequency);
 }
 
 // ---------------------------------------------
